relayburn-sdk: collapse analyze/compare_archive.rs into compare.rs (#347)
The module was a 5-line wrapper around `Ledger::query_turns` +
`build_compare_table` plus ~700 lines of fixture/helper boilerplate to
test what is fundamentally a delegation. Most tests duplicated coverage
already provided by `build_compare_table`'s own tests
(`groups_unclassified_turns_under_unclassified`, the model-filter cases,
the min-sample flagging, etc.) or by the ledger's query tests
(since/project/session filters).

Move `compare_from_archive` + `CompareFromArchiveResult` into
`compare.rs` next to `build_compare_table`; replace the nine
mostly-redundant tests with one focused round-trip smoke test that
appends turns to a real ledger and verifies the wrapper produces the
expected shape. SDK surface unchanged — `compare_from_archive` and
`CompareFromArchiveResult` still re-exported from `relayburn_sdk::analyze`.

Closes #347.
