+++ crates/relayburn-sdk/src/analyze.rs
@@ -23,7 +23,6 @@
 
 pub mod claude_md;
 pub mod compare;
-pub mod compare_archive;
 pub mod cost;
 pub mod fidelity;
 pub mod findings;
@@ -49,10 +48,9 @@ pub use claude_md::{
     TrimRecommendation,
 };
 pub use compare::{
-    build_compare_table, CompareCategory, CompareCell, CompareOptions, CompareTable, CompareTotals,
-    DEFAULT_MIN_SAMPLE,
+    build_compare_table, compare_from_archive, CompareCategory, CompareCell, CompareFromArchiveResult,
+    CompareOptions, CompareTable, CompareTotals, DEFAULT_MIN_SAMPLE,
 };
-pub use compare_archive::{compare_from_archive, CompareFromArchiveResult};
 pub use cost::{
     cost_for_turn, cost_for_usage, lookup_model_rate, sum_costs, CostBreakdown, CostForUsageOptions,
 };
